Get a triangle rendering on screen
This is part of the way through the Hello-Triangle lesson. Still need to
do the Element buffer part and the challenges. But I thought this was a
good stoppig point to get committed since it was a little more
challenging to port over to odin than the last lessons were.